got microtonal scales working, with quantization. still needs UI refinement

+++ src/StepDatum.h
@@ -22,7 +22,7 @@ typedef struct StepDatum
 	uint8_t cv2speed;
 	// utility variables - dont need to be saved.
 	uint8_t noteStatus;		  // if note is playing or not
-	uint8_t notePlaying;	  // stores the note that is played so it can be turned off.
+	uint16_t notePlaying;	 // stores the note that is played so it can be turned off.
 							  //int32_t			offset;		    // note start time offset in mcs from the beat start - recalculated each beat
 	uint16_t arpStatus;		  // if note is playing or not. Value indicates arp number.
 	uint32_t framesRemaining; // number of frames to continue playing.
